update CRUD user

+++ src/main/java/vn/vpgh/jobhunter/controller/UserController.java
@@ -1,6 +1,9 @@
 package vn.vpgh.jobhunter.controller;
 
 import com.turkraft.springfilter.boot.Filter;
+
+import jakarta.validation.Valid;
+
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
@@ -9,12 +12,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import vn.vpgh.jobhunter.domain.User;
+import vn.vpgh.jobhunter.domain.dto.ResCreateUserDTO;
+import vn.vpgh.jobhunter.domain.dto.ResUpdateUserDTO;
+import vn.vpgh.jobhunter.domain.dto.ResUserDTO;
 import vn.vpgh.jobhunter.domain.dto.ResultPaginationDTO;
 import vn.vpgh.jobhunter.service.UserService;
 import vn.vpgh.jobhunter.util.annotation.ApiMessage;
 import vn.vpgh.jobhunter.util.error.IdInvalidException;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v0.1")
@@ -28,34 +32,56 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public ResponseEntity<User> createNewUser(@RequestBody User reqUser) {
+    @ApiMessage("Create a user")
+    public ResponseEntity<ResCreateUserDTO> createNewUser(@Valid @RequestBody User reqUser) throws IdInvalidException {
+        boolean isEmailExist = this.userService.isEmaliExist(reqUser.getEmail());
+        if (isEmailExist) {
+            throw new IdInvalidException("Email already exists");
+        }
         reqUser.setPassword(this.passwordEncoder.encode(reqUser.getPassword()));
-        return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.handleSaveUser(reqUser));
+        User user = this.userService.handleSaveUser(reqUser);
+        return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.convertToResCreateUserDTO(user));
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) {
+    @ApiMessage("Delete a user")
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        User user = this.userService.getUserById(id);
+        if (user == null) {
+            throw new IdInvalidException("There is no user with id of " + id);
+        }
+
         this.userService.deleteUserById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<User> getUser(@PathVariable("id") long id) throws IdInvalidException {
-        if (id < 0) {
-            throw new IdInvalidException("id is less than 0");
+    @ApiMessage("Get a user")
+    public ResponseEntity<ResUserDTO> getUser(@PathVariable("id") long id) throws IdInvalidException {
+        User user = this.userService.getUserById(id);
+        if (user == null) {
+            throw new IdInvalidException("There is no user with id of " + id);
         }
-        return ResponseEntity.status(HttpStatus.OK).body(this.userService.getUserById(id));
+
+        return ResponseEntity.status(HttpStatus.OK).body(this.userService.convertToResUserDTO(user));
     }
 
     @GetMapping("/users")
     @ApiMessage("Get all users")
-    public ResponseEntity<ResultPaginationDTO> getAllUsers(@Filter Specification<User> specification, Pageable pageable) {
+    public ResponseEntity<ResultPaginationDTO> getAllUsers(@Filter Specification<User> specification,
+            Pageable pageable) {
         return ResponseEntity.status(HttpStatus.OK).body(this.userService.getAllUsers(specification, pageable));
     }
 
     @PutMapping("/users")
-    public ResponseEntity<User> updateUser(@RequestBody User reqUser) {
-        return ResponseEntity.status(HttpStatus.OK).body(this.userService.handleUpdateUser(reqUser));
+    @ApiMessage("Update a user")
+    public ResponseEntity<ResUpdateUserDTO> updateUser(@RequestBody User reqUser) throws IdInvalidException {
+        User user = this.userService.handleUpdateUser(reqUser);
+        if (user == null) {
+            throw new IdInvalidException("There is no user with id of " + reqUser.getId());
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(this.userService.convertToResUpdateUserDTO(user));
     }
 
 }
